+++ src/Plugin.php
@@ -3,88 +3,68 @@
 namespace carlcs\redactorcustomstyles;
 
 use carlcs\redactorcustomstyles\assets\customcp\CustomCpAsset;
-use carlcs\redactorcustomstyles\assets\redactorplugin\RedactorPluginAsset;
 use Craft;
-use craft\helpers\Json;
+use craft\helpers\FileHelper;
 use craft\redactor\Field as RedactorField;
 use craft\redactor\events\RegisterPluginPathsEvent;
 use yii\base\Event;
 
 /**
- * @property Settings $settings
- * @method Settings getSettings()
  * @method static Plugin getInstance()
  */
 class Plugin extends \craft\base\Plugin
 {
+    // Properties
+    // =========================================================================
+
+    private ?array $_icons = null;
+
     // Public Methods
     // =========================================================================
 
-    /**
-     * @inheritdoc
-     */
     public function init()
     {
         parent::init();
 
         if (Craft::$app->getRequest()->getIsCpRequest()) {
-            Event::on(RedactorField::class, RedactorField::EVENT_REGISTER_PLUGIN_PATHS, [$this, 'registerRedactorPlugin']);
+            Event::on(RedactorField::class, RedactorField::EVENT_REGISTER_PLUGIN_PATHS, function(RegisterPluginPathsEvent $event) {
+                $event->paths[] = Craft::getAlias('@carlcs/redactorcustomstyles/redactor-plugins');
+            });
 
             $view = Craft::$app->getView();
 
             /** @var CustomCpAsset $customCpAsset */
-            $customCpAsset = $view->registerAssetBundle(CustomCpAsset::class);
-            $view->registerAssetBundle(RedactorPluginAsset::class);
-
-            if (($iconSprite = $customCpAsset->iconSprite) === null) {
-                $iconSprite = '@carlcs/redactorcustomstyles/assets/redactorplugin/dist/icons.svg';
-            }
-
-            $iconSpriteUrl = Craft::$app->getAssetManager()->getPublishedUrl($iconSprite, true);
-            $iconIds = $this->getSymbolIds($iconSprite);
-
-            $view->registerJs('Craft.RedactorCustomStyles = '.Json::encode(compact('iconSpriteUrl', 'iconIds')).';');
+            $view->registerAssetBundle(CustomCpAsset::class);
+            $view->registerJsWithVars(
+                fn($variables) => "Craft.RedactorCustomStyles = $variables",
+                [['icons' => $this->_getIcons()]],
+            );
         }
     }
 
-    /**
-     * Registers the Redactor plugin and its assets.
-     *
-     * @param RegisterPluginPathsEvent $event
-     */
-    public function registerRedactorPlugin(RegisterPluginPathsEvent $event)
-    {
-        $event->paths[] = Craft::getAlias('@carlcs/redactorcustomstyles/assets/redactorplugin/_redactorplugin');
-    }
-
-    // Protected Methods
+    // Private Methods
     // =========================================================================
 
-    /**
-     * Returns the symbol IDs from a SVG sprite.
-     *
-     * @param string $path
-     * @return array
-     */
-    protected function getSymbolIds(string $path): array
+    private function _getIcons(): array
     {
-        $doc = new \DOMDocument();
-        $doc->load(Craft::getAlias($path));
-
-        $ids = [];
-        foreach ($doc->getElementsByTagName('symbol') as $symbol) {
-            /** @var \DOMElement $symbol */
-            $ids[] = $symbol->getAttribute('id');
+        if ($this->_icons !== null) {
+            return $this->_icons;
         }
 
-        return $ids;
-    }
+        $dirs = [
+            Craft::getAlias('@carlcs/redactorcustomstyles/assets/icons'),
+            Craft::getAlias('@config/redactor/resources'),
+        ];
 
-    /**
-     * @inheritdoc
-     */
-    protected function createSettingsModel(): Settings
-    {
-        return new Settings();
+        $this->_icons = [];
+        foreach ($dirs as $dir) {
+            if (is_dir($dir)) {
+                foreach (FileHelper::findFiles($dir, ['only' => ['*.svg']]) as $icon) {
+                    $this->_icons[pathinfo($icon, PATHINFO_FILENAME)] = file_get_contents($icon);
+                }
+            }
+        }
+
+        return $this->_icons;
     }
 }